+++ R/snakeplot-package.R
@@ -2,7 +2,7 @@
 #' @importFrom graphics lines par plot.new plot.window points polygon rect
 #'   segments strwidth text
 #' @importFrom grDevices col2rgb dev.off hsv pdf rgb
-#' @importFrom stats cor median runif weighted.mean
+#' @importFrom stats cor median na.omit runif sd weighted.mean
 "_PACKAGE"
 
 ## usethis namespace: start